correcion faltaba que dgv mostrara precio

+++ TP_WINFORM/Dominio/Articulo.cs
@@ -23,6 +23,7 @@ namespace Dominio
         public Marca Marca { get; set; }
         public Categoria Categoria { get; set; }
         public List<Imagen> Imagen { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
+        
     }
 }
